feat: Created custom answers on forbidden request

+++ library-backend/src/authors/authors.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Author } from './entities/author.entity';
@@ -6,30 +6,48 @@ import { CreateAuthorDto } from './dto/create-author.dto';
 
 @Injectable()
 export class AuthorsService {
-  constructor(
-    @InjectRepository(Author)
-    private authorsRepository: Repository<Author>,
-  ) {}
+    constructor(
+        @InjectRepository(Author)
+        private authorsRepository: Repository<Author>,
+    ) { }
 
-  create(createAuthorDto: CreateAuthorDto): Promise<Author> {
-    const author = this.authorsRepository.create(createAuthorDto);
-    return this.authorsRepository.save(author);
-  }
 
-  findAll(): Promise<Author[]> {
-    return this.authorsRepository.find();
-  }
+    async findOne(id: number): Promise<Author> {
+        const author = await this.authorsRepository.findOneBy({ id });
 
-  findOne(id: number): Promise<Author> {
-    return this.authorsRepository.findOneBy({ id });
-  }
+        if (!author) {
+            throw new NotFoundException(`Author with ID ${id} not found`);
+        }
 
-  async update(id: number, updateAuthorDto: CreateAuthorDto): Promise<Author> {
-    await this.authorsRepository.update(id, updateAuthorDto);
-    return this.findOne(id);
-  }
+        return author;
+    }
 
-  async remove(id: number): Promise<void> {
-    await this.authorsRepository.delete(id);
-  }
+    async update(id: number, updateAuthorDto: CreateAuthorDto): Promise<Author> {
+        const author = await this.authorsRepository.findOneBy({ id });
+
+        if (!author) {
+            throw new NotFoundException(`Author with ID ${id} not found`);
+        }
+
+        await this.authorsRepository.update(id, updateAuthorDto);
+        return this.findOne(id); // Retorna o autor atualizado
+    }
+
+    async remove(id: number): Promise<void> {
+        const author = await this.authorsRepository.findOneBy({ id });
+
+        if (!author) {
+            throw new NotFoundException(`Author with ID ${id} not found`);
+        }
+
+        await this.authorsRepository.delete(id);
+    }
+    create(createAuthorDto: CreateAuthorDto): Promise<Author> {
+        const author = this.authorsRepository.create(createAuthorDto);
+        return this.authorsRepository.save(author);
+    }
+
+    findAll(): Promise<Author[]> {
+        return this.authorsRepository.find();
+    }
 }
